client/requirements.txt
gradio
websocket-client
numpy
librosa
soundfile
